requirements.txt
# Core Django
Django
djangorestframework>=3.15.0,<4.0.0
django-filter>=24.3,<25.0
django-cors-headers>=4.3.1,<5.0.0
django-environ>=0.12.0,<1.0.0

# Database
psycopg2-binary>=2.9.9,<3.0.0

# Authentication & Security
djangorestframework-simplejwt>=5.3.1,<6.0.0

# Channels & WebSockets
channels[daphne]>=4.0.0,<5.0.0
channels-redis>=4.2.0,<5.0.0
daphne>=4.1.0,<5.0.0

# Redis
redis>=5.0.0,<6.0.0
django-redis>=5.4.0,<6.0.0

# Media & Files
Pillow>=10.4.0,<11.0.0

# Utilities
pytz>=2024.1
python-dateutil>=2.9.0,<3.0.0
pyyaml>=6.0.1,<7.0.0
user-agents>=2.2.0,<3.0.0
django-taggit>=5.0.0,<6.0.0
django-multiselectfield>=0.1.12,<1.0.0

# Production Server
gunicorn>=21.2.0,<22.0.0
uvicorn[standard]>=0.30.1,<1.0.0

# Monitoring & Error Tracking
sentry-sdk>=2.10.0,<3.0.0

# Optional: AWS S3 for static/media file storage
# boto3>=1.34.0,<2.0.0
# django-storages>=1.14.0,<2.0.0

# Optional: PostgreSQL connection pooling
# psycogreen>=1.0.2,<2.0.0

# Development dependencies (install with: pip install -r requirements-dev.txt)
# See requirements-dev.txt for testing and development tools
